kubernetes/finance/code/requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.4.0
Flask==1.1.1
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
mysql-connector==2.2.9
openapi-spec-validator==0.2.8
pyrsistent==0.15.5
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
zipp==0.6.0
